refactor(database): separate select type column
