+++ packages/blocks/src/database-block/components/column-type/index.ts
@@ -2,7 +2,7 @@ import { registerColumnRenderer } from '../../register.js';
 import { MultiSelectColumnRenderer } from './multi-select.js';
 import { NumberColumnRenderer } from './number.js';
 import { RichTextColumnRenderer } from './rich-text.js';
-import { SelectColumnRenderer } from './select.js';
+import { SelectColumnRenderer } from './select/select.js';
 
 export function registerInternalRenderer() {
   registerColumnRenderer(NumberColumnRenderer);
