+++ packages/blocks/src/database-block/consts.ts
@@ -6,3 +6,8 @@ export const DEFAULT_COLUMN_MIN_WIDTH = 100;
 export const DEFAULT_COLUMN_TITLE_HEIGHT = 40;
 /** column title height */
 export const DEFAULT_ADD_BUTTON_WIDTH = 40;
+
+/** select input max length */
+export const INPUT_MAX_LENGTH = 10;
+/** edit select popup width */
+export const SELECT_EDIT_POPUP_WIDTH = 345;
